--- a/3rdparty/simdlib/cmake/checks/cpu_sse.cpp
+++ b/3rdparty/simdlib/cmake/checks/cpu_sse.cpp
@@ -1,0 +1,2 @@
+#include <xmmintrin.h>
+int main() { return 0; }