--- a/3rdparty/simdlib/cmake/checks/cpu_sse2.cpp
+++ b/3rdparty/simdlib/cmake/checks/cpu_sse2.cpp
@@ -1,0 +1,2 @@
+#include <emmintrin.h>
+int main() { return 0; }