… .drone.yml
… .travis.yml
… 3rdparty/simdlib/CMakeLists.txt
… 3rdparty/simdlib/README.md
… 3rdparty/simdlib/Simd/SimdAvx2BgrToBgra.cpp
… 3rdparty/simdlib/Simd/SimdAvx2BgrToGray.cpp
… 3rdparty/simdlib/Simd/SimdAvx2BgrToRgb.cpp
… 3rdparty/simdlib/Simd/SimdAvx2BgrToRgba.cpp
… 3rdparty/simdlib/Simd/SimdAvx2BgraToGray.cpp
… 3rdparty/simdlib/Simd/SimdAvx2Deinterleave.cpp
… 3rdparty/simdlib/Simd/SimdAvx2GaussianBlur3x3.cpp
… 3rdparty/simdlib/Simd/SimdAvx2GrayToBgr.cpp
… 3rdparty/simdlib/Simd/SimdAvx2GrayToBgra.cpp
… 3rdparty/simdlib/Simd/SimdAvx2Interleave.cpp
… 3rdparty/simdlib/Simd/SimdAvx2Operation.cpp
… 3rdparty/simdlib/Simd/SimdAvx2Reduce.cpp
… 3rdparty/simdlib/Simd/SimdAvx2ReduceGray2x2.cpp
… 3rdparty/simdlib/Simd/SimdAvx2ReduceGray3x3.cpp
… 3rdparty/simdlib/Simd/SimdAvx2ReduceGray4x4.cpp
… 3rdparty/simdlib/Simd/SimdAvx2ReduceGray5x5.cpp
… 3rdparty/simdlib/Simd/SimdAvx2ResizeBilinear.cpp
… 3rdparty/simdlib/Simd/SimdAvx2Resizer.cpp
… 3rdparty/simdlib/Simd/SimdAvx2RgbToGray.cpp
… 3rdparty/simdlib/Simd/SimdAvx2RgbaToGray.cpp
… 3rdparty/simdlib/Simd/SimdAvx2StretchGray2x2.cpp
… 3rdparty/simdlib/Simd/SimdBase.h
… 3rdparty/simdlib/Simd/SimdBaseBgrToBgra.cpp
… 3rdparty/simdlib/Simd/SimdBaseBgrToRgb.cpp
… 3rdparty/simdlib/Simd/SimdBaseBgrToRgba.cpp
… 3rdparty/simdlib/Simd/SimdBaseBgraToBgr.cpp