Revamp of #657 to have a clean history.
Now that the release is over, we should push to be able to merge it. This way, we would be able to fix build issues later.

TODO:
- [x] fix CI build issue
- [x] try build and run on ARM platforms
- [x] check some test and perf files

---

Main changes:

- I have removed Haar cascade detection that is useless nowadays and I was having some issues on different platforms
- try to trimmed the most possible the lib to avoid having huge number of files, that makes it unmaintainable

---

simdlib cannot be disabled since the idea is to have code optimized using SIMD intrinsics on x86 and ARM platforms. It contains fallback code to plain C++ code for the other platforms, or when the compiler does not support the corresponding flags.
So if for whatever reason the optimized code cannot be compiled, the `base` C++ code must be compiled instead.

For now, there are few places in ViSP code that contains intrinsics, but I think ultimately and ideally we should move them to dedicated files. This is clearly not the priority.