@s-trinh 2020-06-07
This PR is ready. All builds are green.
Probably compiler warnings will need to be fixed later. Also, testing that `pkg-config` still works.

Below some performance numbers. `master` is commit `4905ef68c`. `simdlib` is this PR.

---

## i5-8500 CPU @ 3.00GHz

### perfColorConversion

- image size (Klimt): `558x560`
- `simdlib nThreads=1`

| Benchmark rgba to grayscale (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (naive code) | 446.55 | 442.88 | 1.01 |

| Benchmark rgba to grayscale (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (ViSP) | 84.43 | 60.25 | 1.40 |

| Benchmark grayscale to rgba (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (naive code) | 114.34 | 112.65 | 1.02 |

| Benchmark grayscale to rgba (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (ViSP) | 117.17 | 107.07 | 1.09 |

| Benchmark split RGBa (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark split RGBa (ViSP) | 380.46 | 66.04 | 5.76 |

| Benchmark merge to RGBa (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark merge to RGBa (ViSP) | 699.97 | 124.49 | 5.62 |

| Benchmark bgr to grayscale (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (naive code) | 447.16 | 447.06 | 1.00 |

| Benchmark bgr to grayscale (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (ViSP) | 80.06 | 71.27 | 1.12 |
| Benchmark bgr to grayscale (ViSP + OpenCV Mat type) | 81.06 | 71.55 | 1.13 |

| Benchmark bgr to grayscale (OpenCV) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (OpenCV) | 27.24 | 27.18 | 1.00 |

| Benchmark bgr to rgba (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (naive code) | 165.25 | 164.77 | 1.00 |

| Benchmark bgr to rgba (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (ViSP) | 165.84 | 114.28 | 1.45 |
| Benchmark bgr to rgba (ViSP + OpenCV Mat type) | 113.91 | 115.24 | 0.99 |

For some functions, `nThreads` can be passed as a parameter for multithreading using OpenMP. This is an explicit parameter to allow two different "multithreading approaches":

- "multithreading directly in the algorithm", for instance for color conversion: the workload can be splitted across the image size (`nThreads` can be `0` to use all the CPU threads or the number of CPU threads)
- "application multithreading", for instance for batch processing: color conversion to process 4 images in parallel (`nThreads` must be `1` and a pool of threads can be created to process multiple images)

Here, benchmark with `nThreads=1` shows that the simdlib implementation is better compared to the previous custom SSE implementation for color conversion:

- probably because it uses AVX2 intrinsics (vectors are 256 bit-size) while the master uses SSE intrinsics
- maybe a better SIMD implementation, maybe better cache locality

---

- `simdlib nThreads=6`

| Benchmark rgba to grayscale (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (naive code) | 446.55 | 445.51 | 1.00 |

| Benchmark rgba to grayscale (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (ViSP) | 84.43 | 26.48 | 3.19 |

| Benchmark grayscale to rgba (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (naive code) | 114.34 | 111.99 | 1.02 |

| Benchmark grayscale to rgba (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (ViSP) | 117.17 | 108.02 | 1.08 |

| Benchmark split RGBa (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark split RGBa (ViSP) | 380.46 | 66.36 | 5.73 |

| Benchmark merge to RGBa (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark merge to RGBa (ViSP) | 699.97 | 126.89 | 5.52 |

| Benchmark bgr to grayscale (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (naive code) | 447.16 | 447.64 | 1.00 |

| Benchmark bgr to grayscale (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (ViSP) | 80.06 | 27.80 | 2.88 |
| Benchmark bgr to grayscale (ViSP + OpenCV Mat type) | 81.06 | 28.36 | 2.86 |

| Benchmark bgr to grayscale (OpenCV) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (OpenCV) | 27.24 | 27.67 | 0.98 |

| Benchmark bgr to rgba (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (naive code) | 165.25 | 170.32 | 0.97 |

| Benchmark bgr to rgba (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (ViSP) | 165.84 | 119.64 | 1.39 |
| Benchmark bgr to rgba (ViSP + OpenCV Mat type) | 113.91 | 120.20 | 0.95 |

Better performance for color conversion when using multithreading.

- image size: `1920x1080`

| Benchmark rgba to grayscale (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (naive code) | 3199.13 | 3091.02 | 1.03 |

| Benchmark rgba to grayscale (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (ViSP) | 738.83 | 600.62 | 1.23 |

| Benchmark grayscale to rgba (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (naive code) | 1459.50 | 1439.74 | 1.01 |

| Benchmark grayscale to rgba (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (ViSP) | 1456.15 | 1432.52 | 1.02 |

| Benchmark split RGBa (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark split RGBa (ViSP) | 3183.73 | 835.79 | 3.81 |

| Benchmark merge to RGBa (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark merge to RGBa (ViSP) | 5476.73 | 1876.57 | 2.92 |

| Benchmark bgr to grayscale (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (naive code) | 2997.65 | 2998.01 | 1.00 |

| Benchmark bgr to grayscale (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (ViSP) | 645.20 | 568.01 | 1.14 |
| Benchmark bgr to grayscale (ViSP + OpenCV Mat type) | 649.94 | 532.93 | 1.22 |

| Benchmark bgr to grayscale (OpenCV) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (OpenCV) | 151.70 | 149.63 | 1.01 |

| Benchmark bgr to rgba (naive code) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (naive code) | 1925.89 | 1910.07 | 1.01 |

| Benchmark bgr to rgba (ViSP) - time unit: us | master | simdlib nThreads=1 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (ViSP) | 1926.25 | 1714.95 | 1.12 |
| Benchmark bgr to rgba (ViSP + OpenCV Mat type) | 1743.13 | 1717.91 | 1.01 |

---

| Benchmark rgba to grayscale (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (naive code) | 3199.13 | 3093.37 | 1.03 |

| Benchmark rgba to grayscale (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark rgba to grayscale (ViSP) | 738.83 | 343.94 | 2.15 |

| Benchmark grayscale to rgba (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (naive code) | 1459.50 | 1442.22 | 1.01 |

| Benchmark grayscale to rgba (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark grayscale to rgba (ViSP) | 1456.15 | 1442.00 | 1.01 |

| Benchmark split RGBa (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark split RGBa (ViSP) | 3183.73 | 826.39 | 3.85 |

| Benchmark merge to RGBa (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark merge to RGBa (ViSP) | 5476.73 | 1880.39 | 2.91 |

| Benchmark bgr to grayscale (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (naive code) | 2997.65 | 2995.65 | 1.00 |

| Benchmark bgr to grayscale (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (ViSP) | 645.20 | 228.21 | 2.83 |
| Benchmark bgr to grayscale (ViSP + OpenCV Mat type) | 649.94 | 230.35 | 2.82 |

| Benchmark bgr to grayscale (OpenCV) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to grayscale (OpenCV) | 151.70 | 150.37 | 1.01 |

| Benchmark bgr to rgba (naive code) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (naive code) | 1925.89 | 1917.24 | 1.00 |

| Benchmark bgr to rgba (ViSP) - time unit: us | master | simdlib nThreads=6 | Speed-up |
| --- | --- | --- | --- |
| Benchmark bgr to rgba (ViSP) | 1926.25 | 1730.93 | 1.11 |
| Benchmark bgr to rgba (ViSP + OpenCV Mat type) | 1743.13 | 1725.50 | 1.01 |

No performance issue with HD images.

### perfImageResize

- Klimt: resize from `558x560` to `293x137`

| Nearest Neighbor image resize (naive code) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (naive code) | 206.31 | 342.25 | 0.60 |
| Benchmark Nearest Neighbor RGBa image resize (naive code) | 286.45 | 430.32 | 0.67 |

| Nearest Neighbor image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (ViSP) (1 thread) | 147.60 | 371.16 | 0.40 |
| Benchmark Nearest Neighbor uchar image resize (ViSP) (6 threads) | 30.09 | 69.58 | 0.43 |
| Benchmark Nearest Neighbor RGBa image resize (ViSP) (1 thread) | 227.74 | 452.76 | 0.50 |
| Benchmark Nearest Neighbor RGBa image resize (ViSP) (6 threads) | 50.84 | 91.65 | 0.55 |

| Bilinear image resize (naive code) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (naive code) | 429.60 | 437.47 | 0.98 |
| Benchmark Bilinear RGBa image resize (naive code) | 1002.37 | 1007.92 | 0.99 |

| Bilinear image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (ViSP) | 41.97 | 13.85 | 3.03 |
| Benchmark Bilinear RGBa image resize (ViSP) | 128.53 | 69.42 | 1.85 |

| Bicubic image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bicubic uchar image resize (ViSP) (1 thread) | 3138.28 | 6161.67 | 0.51 |
| Benchmark Bicubic uchar image resize (ViSP) (6 threads) | 551.42 | 1089.38 | 0.51 |
| Benchmark Bicubic RGBa image resize (ViSP) (1 thread) | 5880.07 | 9280.68 | 0.63 |
| Benchmark Bicubic RGBa image resize (ViSP) (6 threads) | 1046.80 | 1649.33 | 0.63 |

| Nearest Neighbor image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (OpenCV) | 17.61 | 17.83 | 0.99 |
| Benchmark Nearest Neighbor BGR image resize (OpenCV) | 43.03 | 41.89 | 1.03 |

| Bilinear image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (OpenCV) | 47.25 | 46.19 | 1.02 |
| Benchmark Bilinear BGR image resize (OpenCV) | 113.97 | 116.78 | 0.98 |

| Bicubic image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bicubic uchar image resize (OpenCV) | 128.92 | 129.78 | 0.99 |
| Benchmark Bicubic BGR image resize (OpenCV) | 91.03 | 90.81 | 1.00 |

Now, the pixel coordinates are uniformise as the following:

- `u_dst = (u_src + 0.5) x scale_x - 0.5`
- `v_dst = (v_src + 0.5) x scale_y - 0.5`

Pixel center is now considerate.
This explains the performance degradation for Nearest Neighbor and Bicubic image resize.
But what is important is that the **Bilinear method is faster than the Nearest Neighbor method**. So, Bilinear method should be used for image resizing. `simdlib` is used for Bilinear image resize.

- upscale: resize from `640x480` to `1280x960` (2x)

| Nearest Neighbor image resize (naive code) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (naive code) | 6301.38 | 10478.40 | 0.60 |
| Benchmark Nearest Neighbor RGBa image resize (naive code) | 9107.46 | 13427.02 | 0.68 |

| Nearest Neighbor image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (ViSP) (1 thread) | 4504.77 | 11385.58 | 0.40 |
| Benchmark Nearest Neighbor uchar image resize (ViSP) (6 threads) | 839.31 | 2029.33 | 0.41 |
| Benchmark Nearest Neighbor RGBa image resize (ViSP) (1 thread) | 6959.97 | 14061.58 | 0.49 |
| Benchmark Nearest Neighbor RGBa image resize (ViSP) (6 threads) | 1673.39 | 2870.33 | 0.58 |

| Bilinear image resize (naive code) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (naive code) | 14485.32 | 14690.22 | 0.99 |
| Benchmark Bilinear RGBa image resize (naive code) | 32015.97 | 31985.78 | 1.00 |

| Bilinear image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (ViSP) | 1174.20 | 150.94 | 7.78 |
| Benchmark Bilinear RGBa image resize (ViSP) | 4041.36 | 1136.19 | 3.56 |

| Bicubic image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bicubic uchar image resize (ViSP) (1 thread) | 95611.73 | 189062.40 | 0.51 |
| Benchmark Bicubic uchar image resize (ViSP) (6 threads) | 16703.26 | 32992.19 | 0.51 |
| Benchmark Bicubic RGBa image resize (ViSP) (1 thread) | 177217.07 | 281019.10 | 0.63 |
| Benchmark Bicubic RGBa image resize (ViSP) (6 threads) | 31336.08 | 49549.18 | 0.63 |

| Nearest Neighbor image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (OpenCV) | 108.31 | 108.78 | 1.00 |
| Benchmark Nearest Neighbor BGR image resize (OpenCV) | 273.20 | 265.77 | 1.03 |

| Bilinear image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (OpenCV) | 128.99 | 129.30 | 1.00 |
| Benchmark Bilinear BGR image resize (OpenCV) | 367.94 | 367.07 | 1.00 |

| Bicubic image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bicubic uchar image resize (OpenCV) | 225.78 | 225.93 | 1.00 |
| Benchmark Bicubic BGR image resize (OpenCV) | 636.99 | 637.53 | 1.00 |

- downscale from `1920x1080` to `640x360` (3x)

| Nearest Neighbor image resize (naive code) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (naive code) | 942.01 | 1808.17 | 0.52 |
| Benchmark Nearest Neighbor RGBa image resize (naive code) | 1501.29 | 2328.86 | 0.64 |

| Nearest Neighbor image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (ViSP) (1 thread) | 860.03 | 1943.30 | 0.44 |
| Benchmark Nearest Neighbor uchar image resize (ViSP) (6 threads) | 160.14 | 359.76 | 0.45 |
| Benchmark Nearest Neighbor RGBa image resize (ViSP) (1 thread) | 1302.21 | 2534.21 | 0.51 |
| Benchmark Nearest Neighbor RGBa image resize (ViSP) (6 threads) | 282.28 | 493.12 | 0.57 |

| Bilinear image resize (naive code) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (naive code) | 2190.97 | 2223.88 | 0.99 |
| Benchmark Bilinear RGBa image resize (naive code) | 4532.38 | 4520.22 | 1.00 |

| Bilinear image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (ViSP) | 225.11 | 90.16 | 2.50 |
| Benchmark Bilinear RGBa image resize (ViSP) | 745.77 | 464.62 | 1.61 |

| Bicubic image resize (ViSP) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bicubic uchar image resize (ViSP) (1 thread) | 17905.59 | 32608.79 | 0.55 |
| Benchmark Bicubic uchar image resize (ViSP) (6 threads) | 3140.05 | 5657.53 | 0.56 |
| Benchmark Bicubic RGBa image resize (ViSP) (1 thread) | 33322.69 | 48769.91 | 0.68 |
| Benchmark Bicubic RGBa image resize (ViSP) (6 threads) | 5874.99 | 8554.28 | 0.69 |

| Nearest Neighbor image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Nearest Neighbor uchar image resize (OpenCV) | 28.68 | 29.04 | 0.99 |
| Benchmark Nearest Neighbor BGR image resize (OpenCV) | 65.89 | 66.02 | 1.00 |

| Bilinear image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bilinear uchar image resize (OpenCV) | 70.28 | 70.26 | 1.00 |
| Benchmark Bilinear BGR image resize (OpenCV) | 177.44 | 178.36 | 0.99 |

| Bicubic image resize (OpenCV) - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark Bicubic uchar image resize (OpenCV) | 114.61 | 114.98 | 1.00 |
| Benchmark Bicubic BGR image resize (OpenCV) | 347.38 | 347.26 | 1.00 |

`simdlib` is used for `vpImageTools::INTERPOLATION_AREA` and `vpImageTools::INTERPOLATION_LINEAR`. For now, multithreading is not used. If further tests show that it improves the perf, it can be added later.

### perfImageAddSub

- image size (Klimt): `558x560`

| Benchmark vpImageTools::imageAdd() - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark naive imageAdd() code without saturation | 206.35 | 208.10 | 0.99 |
| Benchmark ViSP imageAdd() code without saturation | 27.98 | 25.30 | 1.11 |
| Benchmark naive imageAdd() code with saturation | 247.17 | 242.62 | 1.02 |
| Benchmark ViSP imageAdd() code with saturation | 26.89 | 24.53 | 1.10 |

| Benchmark vpImageTools::imageSubtract() - time unit: us | master | simdlib | Speed-up |
| --- | --- | --- | --- |
| Benchmark naive imageSub() code without saturation | 210.01 | 208.63 | 1.01 |
| Benchmark ViSP imageSub() code without saturation | 26.69 | 24.36 | 1.10 |
…